fix(test_modeling_rules): move get_containing_pack to utils
