+++ demisto_sdk/utils/utils.py
@@ -1,5 +1,31 @@
 import os
 from configparser import ConfigParser, MissingSectionHeaderError
+from typing import Union
+
+from demisto_sdk.commands.common.content.objects.pack_objects.abstract_pack_objects.json_content_object import \
+    JSONContentObject
+from demisto_sdk.commands.common.content.objects.pack_objects.abstract_pack_objects.yaml_content_object import \
+    YAMLContentObject
+from demisto_sdk.commands.common.content.objects.pack_objects.abstract_pack_objects.yaml_unify_content_object import \
+    YAMLContentUnifiedObject
+from demisto_sdk.commands.common.content.objects.pack_objects.pack import Pack
+
+ContentEntity = Union[YAMLContentUnifiedObject, YAMLContentObject, JSONContentObject]
+
+
+def get_containing_pack(content_entity: ContentEntity) -> Pack:
+    """Get pack object that contains the content entity.
+
+    Args:
+        content_entity: Content entity object.
+
+    Returns:
+        Pack: Pack object that contains the content entity.
+    """
+    pack_path = content_entity.path
+    while pack_path.parent.name.casefold() != 'packs':
+        pack_path = pack_path.parent
+    return Pack(pack_path)
 
 
 def check_configuration_file(command, args):
